feat: 04 inital version

+++ uni-hw-02/Homework-2/05/05.cpp
@@ -6,7 +6,7 @@ const int X = 0;
 const int Y = 7;
 const int Z = 1;
 
-int T(int n) // T(10) = 282 475 249
+long long T(int n) // T(10) = 282 475 249
 {
 	if (n == 0)
 	{
